simplified/unified package structure
