+++ opentelemetry-example/src/main/scala/zio/telemetry/opentelemetry/example/config/package.scala
@@ -1,0 +1,9 @@
+package zio.telemetry.opentelemetry.example
+
+import zio.Has
+
+package object config {
+
+  type Configuration = Has[Configuration.Service]
+
+}
