--- opentelemetry-example/src/main/scala/zio/telemetry/opentelemetry/tracing/example/config/Configuration.scala
@@ -1,24 +1,0 @@
-package zio.telemetry.opentelemetry.tracing.example.config
-
-import pureconfig.ConfigSource
-import pureconfig.generic.auto._
-import zio.Task
-import zio.ZLayer
-import zio.ZIO
-
-object Configuration {
-
-  trait Service {
-    val load: Task[Config]
-  }
-
-  object Live extends Service {
-    val load: Task[Config] =
-      Task.effect(ConfigSource.default.loadOrThrow[Config])
-  }
-
-  val live: ZLayer[Any, Throwable, Configuration] = ZLayer.succeed(Live)
-
-  val load: ZIO[Configuration, Throwable, Config] = ZIO.accessM[Configuration] { _.get.load }
-
-}
