--- opentelemetry/src/main/scala/zio/opentelemetry/tracing/attributevalue/AttributeValueConverter.scala
@@ -1,11 +1,0 @@
-package zio.opentelemetry.tracing.attributevalue
-
-import io.opentelemetry.common.AttributeValue
-
-trait AttributeValueConverter[A] {
-  def convert(value: A): AttributeValue
-}
-
-object AttributeValueConverter {
-  def toAttributeValue[A](value: A)(implicit c: AttributeValueConverter[A]): AttributeValue = c.convert(value)
-}
