--- opentelemetry/src/main/scala/zio/opentelemetry/tracing/package.scala
@@ -1,7 +1,0 @@
-package zio.opentelemetry
-
-import zio.Has
-
-package object tracing {
-  type Tracing = Has[Tracing.Service]
-}
