+++ opentelemetry/src/main/scala/zio/telemetry/opentelemetry/attributevalue/AttributeValueConverterInstances.scala
@@ -1,0 +1,11 @@
+package zio.telemetry.opentelemetry.attributevalue
+
+import io.opentelemetry.common.AttributeValue
+
+object AttributeValueConverterInstances {
+  implicit val stringConverter: AttributeValueConverter[String]   = AttributeValue.stringAttributeValue(_)
+  implicit val booleanConverter: AttributeValueConverter[Boolean] = AttributeValue.booleanAttributeValue(_)
+  implicit val longConverter: AttributeValueConverter[Long]       = AttributeValue.longAttributeValue(_)
+  implicit val doubleConverter: AttributeValueConverter[Double]   = AttributeValue.doubleAttributeValue(_)
+  implicit val intConverter: AttributeValueConverter[Int]         = (value: Int) => longConverter.convert(value.toLong)
+}
