+++ opentelemetry/src/main/scala/zio/telemetry/opentelemetry/package.scala
@@ -1,0 +1,7 @@
+package zio.telemetry
+
+import zio.Has
+
+package object opentelemetry {
+  type Tracing = Has[Tracing.Service]
+}
